Small and timed small moved to namespaces
