+++ Core/Battle/Menu/IGMData/IGMData_PartyItems.cs
@@ -38,9 +38,9 @@ namespace OpenVIII
                     ITEM[0, 1] = new IGMDataItem.Box(null, new Rectangle(SIZE[0].X + 140, SIZE[0].Y + 189, 475, 78), Icons.ID.ITEM, options: Box_Options.Middle); // item name
                     ITEM[0, 2] = new IGMDataItem.Box(null, new Rectangle(SIZE[0].X + 615, SIZE[0].Y + 189, 125, 78), Icons.ID.NUM_, options: Box_Options.Middle | Box_Options.Center); // item count
                     ITEM[0, 3] = new IGMDataItem.Box(null, new Rectangle(SIZE[0].X, SIZE[0].Y + 444, SIZE[0].Width, 78), Icons.ID.HELP, options: Box_Options.Middle); // item description
-                    ITEM[0, 4] = new IGMData_SmallMsgBox(null, SIZE[0].X + 232, SIZE[0].Y + 315, Icons.ID.NOTICE, Box_Options.Center | Box_Options.Middle, SIZE[0]); // Couldn't find any items
-                    ITEM[0, 5] = new IGMData_SmallMsgBox(null, SIZE[0].X + 230, SIZE[0].Y + 291, Icons.ID.NOTICE, Box_Options.Center, SIZE[0]); // over 100 discarded
-                    ITEM[0, 6] = new IGMData_SmallMsgBox(null, SIZE[0].X + 232, SIZE[0].Y + 315, Icons.ID.NOTICE, Box_Options.Center, SIZE[0]); // Recieved item
+                    ITEM[0, 4] = new IGMData.Dialog.Small(null, SIZE[0].X + 232, SIZE[0].Y + 315, Icons.ID.NOTICE, Box_Options.Center | Box_Options.Middle, SIZE[0]); // Couldn't find any items
+                    ITEM[0, 5] = new IGMData.Dialog.Small(null, SIZE[0].X + 230, SIZE[0].Y + 291, Icons.ID.NOTICE, Box_Options.Center, SIZE[0]); // over 100 discarded
+                    ITEM[0, 6] = new IGMData.Dialog.Small(null, SIZE[0].X + 232, SIZE[0].Y + 315, Icons.ID.NOTICE, Box_Options.Center, SIZE[0]); // Recieved item
                     Cursor_Status |= (Cursor_Status.Hidden | (Cursor_Status.Enabled | Cursor_Status.Static));
                 }
 
@@ -131,8 +131,8 @@ namespace OpenVIII
                         ((IGMDataItem.Box)ITEM[0, 1]).Data = Item.DATA?.Name;
                         ((IGMDataItem.Box)ITEM[0, 2]).Data = $"{Item.QTY}";
                         ((IGMDataItem.Box)ITEM[0, 3]).Data = Item.DATA?.Description;
-                        ((IGMData_SmallMsgBox)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
-                        ((IGMData_SmallMsgBox)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
+                        ((IGMData.Dialog.Small)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
+                        ((IGMData.Dialog.Small)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
                         ITEM[0, 1].Show();
                         ITEM[0, 2].Show();
                         ITEM[0, 3].Show();
@@ -153,8 +153,8 @@ namespace OpenVIII
                         //TODO grab card name from start of string
                         ((IGMDataItem.Box)ITEM[0, 2]).Data = $"{card.Value}";
                         ((IGMDataItem.Box)ITEM[0, 3]).Data = "";
-                        ((IGMData_SmallMsgBox)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
-                        ((IGMData_SmallMsgBox)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
+                        ((IGMData.Dialog.Small)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
+                        ((IGMData.Dialog.Small)ITEM[0, 5]).Data = str_Over100.Clone().Replace(DialogSelectedItem, Item.DATA?.Name);
                         ITEM[0, 1].Show();
                         ITEM[0, 2].Show();
                         ITEM[0, 3].Hide();
